+++ tests/src/IntegrationTest/ObjectMapperTest.php
@@ -24,6 +24,8 @@ use Rekalogika\Mapper\Tests\Fixtures\ObjectMapper\MoneyDtoForProxy;
 use Rekalogika\Mapper\Tests\Fixtures\ObjectMapper\MoneyDtoForTargetInvalidTypeHint;
 use Rekalogika\Mapper\Tests\Fixtures\ObjectMapper\MoneyDtoForTargetModification;
 use Rekalogika\Mapper\Tests\Fixtures\ObjectMapper\MoneyDtoForTargetReplacement;
+use Rekalogika\Mapper\Tests\Fixtures\ObjectMapper\MoneyDtoInterface;
+use Rekalogika\Mapper\Tests\Fixtures\ObjectMapper\MoneyDtoToo;
 use Rekalogika\Mapper\Tests\Fixtures\ObjectMapper\Person;
 use Rekalogika\Mapper\Tests\Fixtures\ObjectMapper\PersonDto;
 use Rekalogika\Mapper\Tests\Services\ObjectMapper\MoneyObjectMapper;
@@ -54,6 +56,27 @@ class ObjectMapperTest extends FrameworkTestCase
     public function testMoneyDtoToMoney(): void
     {
         $moneyDto = new MoneyDto('100.00', 'USD');
+        $result = $this->mapper->map($moneyDto, Money::class);
+
+        $this->assertInstanceOf(Money::class, $result);
+        $this->assertSame('100.00', $result->getAmount()->__toString());
+        $this->assertSame('USD', $result->getCurrency()->getCurrencyCode());
+    }
+
+    public function testMoneyToMoneyDtoInterface(): void
+    {
+        $money = Money::of('100.00', 'USD');
+        $result = $this->mapper->map($money, MoneyDtoInterface::class);
+
+        $this->assertInstanceOf(MoneyDtoInterface::class, $result);
+        $this->assertInstanceOf(MoneyDto::class, $result);
+        $this->assertSame('100.00', $result->getAmount());
+        $this->assertSame('USD', $result->getCurrency());
+    }
+
+    public function testMoneyDtoInterfaceToMoney(): void
+    {
+        $moneyDto = new MoneyDtoToo('100.00', 'USD');
         $result = $this->mapper->map($moneyDto, Money::class);
 
         $this->assertInstanceOf(Money::class, $result);